requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
astroid==4.0.1
async-timeout==5.0.1
billiard==4.2.1
black==23.7.0
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.0
dill==0.4.0
dj-database-url==2.1.0
Django==5.2.5
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.122.0
flake8==6.0.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.3.0
isort==7.0.0
kombu==5.3.4
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
playwright==1.57.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
pycodestyle==2.10.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.0.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
redis==5.0.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.50.0
stripe==13.0.1
tabulate==0.9.0
tomli==2.3.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.13
